requirements.txt
python-telegram-bot
telegram~=0.0.1
python-jenkins
xmltodict